# openapi.yaml
openapi: 3.0.0
paths:
...
  /countries:
    get:
      operationId: CountriesController_list
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: &a7
        - bearer: []
      tags: &a8
        - countries
    post:
      operationId: CountriesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCountryDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: *a7
      tags: *a8
  /countries/import:
    post:
      operationId: CountriesController_importCsv
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: *a7
      tags: *a8
...
  /audit/project/{projectId}:
    get:
      operationId: AuditController_listProject
      parameters:
        - name: projectId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security:
        - bearer: []
      tags:
        - audit
...
  /render-profiles:
    get:
      operationId: RenderController_listProfiles
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: &a19
        - bearer: []
      tags: &a20
        - render
  /render-profiles/{id}:
    patch:
      operationId: RenderController_updateProfile
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRenderProfileDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: *a19
      tags: *a20
  /render-batches:
    post:
      operationId: RenderController_createBatch
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRenderBatchDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: *a19
      tags: *a20
  /render-batches/{id}:
    get:
      operationId: RenderController_getBatch
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: *a19
      tags: *a20
  /render-batches/{id}/cancel:
    post:
      operationId: RenderController_cancelBatch
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: *a19
      tags: *a20
  /downloads:
    post:
      operationId: DownloadsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDownloadDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: &a21
        - bearer: []
      tags: &a22
        - downloads
  /downloads/{id}:
    get:
      operationId: DownloadsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: *a21
      tags: *a22
  /downloads/{id}/cancel:
    post:
      operationId: DownloadsController_cancel
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: *a21
      tags: *a22
...
info:
  title: AG Go API
  description: AG Go backend API
  version: 0.1.0
  contact: {}
...
servers:
  - url: http://localhost:3737/api
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateCountryDto:
      type: object
      properties: {}
...
    UpdateRenderProfileDto:
      type: object
      properties: {}
    CreateRenderBatchDto:
      type: object
      properties: {}
    CreateDownloadDto:
      type: object
      properties: {}
...
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    ApiResponse:
      type: object
      required:
        - data
        - requestId
        - timestamp
        - success
        - error
      properties:
        data: {}
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        success:
          type: boolean
        error:
          allOf:
            - $ref: "#/components/schemas/ApiError"
          nullable: true
...